adds weapon class and base methods to the thief

+++ src/main/java/Fighter.java
@@ -2,7 +2,7 @@ import java.util.ArrayList;
 
 public class Fighter extends Character {
 
-    public Fighter(String name, int HP, int damage) {
+    public Fighter(String name, double HP, double damage) {
         super(name, HP, damage);
         super.weapons = new ArrayList<Weapon>();
         super.loot = new ArrayList<Treasure>();
